Extract Tree List and Array
