+++ README.md
@@ -1,1 +1,12 @@
 # leetcode-utils
+
+## Array
+
+- swap
+
+## Tree
+
+- treeToArray
+- arrayToTree
+
+
