+++ src/List.js
@@ -1,0 +1,10 @@
+class ListNode {
+  constructor(val) {
+    this.val = val;
+    this.next = null;
+  }
+}
+
+module.exports = {
+  ListNode,
+};
